requirements.txt
requests
python-dotenv
pytest
pandas
sklearn
flask